Impl Clone and round-trip conversion for Fillet and FilletIntoIter.
`FilletIntoIter` owns the same shape of heap allocation as `Fillet`
so conversion can often be as simple as taking ownership.
When elements have been removed from `FilletIntoIter`, the initialized
portion is moved to the beginning of the allocation, and it is
reallocated, which sometimes will cause the allocation to be copied.